src/test/java/edu/sinclair/ssp/service/reference/FundingSourceServiceTest.java
package edu.sinclair.ssp.service.reference;

import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import java.util.List;
import java.util.UUID;

import org.junit.Before;
import org.junit.Test;

import edu.sinclair.ssp.model.reference.FundingSource;

public class FundingSourceServiceTest {

	private FundingSourceService service;
	
	@Before
	public void setup(){
		service = new FundingSourceService();
	}
	
	@Test
	public void testGetAll() {
		List<FundingSource> all = service.getAll();
		assertTrue(all.size()>0);
	}

	@Test
	public void testGet() {
		assertNotNull(service.get(UUID.randomUUID()));
	}

	@Test
	public void testSave() {
		assertNotNull(service.save(new FundingSource(UUID.randomUUID(), "test source")));
	}

	@Test
	public void testDelete() {
		service.delete(UUID.randomUUID());
		assertTrue(true);
	}

}
